+++ src/frontend/apps/lti_site/components/PublicVideoDashboard/PublicVODDashboard/index.tsx
@@ -2,13 +2,12 @@ import { Box } from 'grommet';
 import React from 'react';
 import { Redirect } from 'react-router-dom';
 
-import { DownloadVideo } from 'components/DownloadVideo';
 import { FULL_SCREEN_ERROR_ROUTE } from 'components/ErrorComponents/route';
-import { Transcripts } from 'components/Transcripts';
 import VideoPlayer from 'components/VideoPlayer';
 import { useCurrentVideo } from 'data/stores/useCurrentRessource/useCurrentVideo';
 import { useTimedTextTrack } from 'data/stores/useTimedTextTrack';
-import { timedTextMode, TimedTextTranscript, uploadState } from 'types/tracks';
+import { uploadState } from 'types/tracks';
+import { VideoWidgetProvider } from 'components/VideoWidgetProvider';
 
 interface PublicVODDashboardProps {
   playerType: string;
@@ -20,14 +19,6 @@ export const PublicVODDashboard = ({ playerType }: PublicVODDashboardProps) => {
     state.getTimedTextTracks(),
   );
 
-  const transcripts = timedTextTracks
-    .filter((track) => track.is_ready_to_show)
-    .filter((track) =>
-      video.has_transcript === false && video.should_use_subtitle_as_transcript
-        ? timedTextMode.SUBTITLE === track.mode
-        : timedTextMode.TRANSCRIPT === track.mode,
-    );
-
   if (video.upload_state === uploadState.DELETED) {
     return <Redirect push to={FULL_SCREEN_ERROR_ROUTE('videoDeleted')} />;
   }
@@ -37,16 +28,14 @@ export const PublicVODDashboard = ({ playerType }: PublicVODDashboardProps) => {
   }
 
   return (
-    <Box>
+    <Box background={{ color: 'bg-marsha' }}>
       <VideoPlayer
         video={video}
         playerType={playerType}
         timedTextTracks={timedTextTracks}
       />
-      {video.show_download && <DownloadVideo urls={video.urls} />}
-      {transcripts.length > 0 && (
-        <Transcripts transcripts={transcripts as TimedTextTranscript[]} />
-      )}
+
+      <VideoWidgetProvider isLive={false} isTeacher={false} />
     </Box>
   );
 };